fix: patch

+++ erpnext/patches/v12_0/rename_pos_closing_doctype.py
@@ -14,6 +14,7 @@ def execute():
 		frappe.rename_doc('DocType', 'POS Closing Voucher Details', 'POS Closing Entry Details', force=True)
 		frappe.reload_doc('Selling', 'doctype', 'POS Closing Entry Details')
 
+	if frappe.db.exists("DocType", "POS Closing Voucher"):
 		frappe.delete_doc("DocType", "POS Closing Voucher")
 		frappe.delete_doc("DocType", "POS Closing Voucher Taxes")
 		frappe.delete_doc("DocType", "POS Closing Voucher Details")
